Revert "linux: 4.19.181"
This reverts commit 8921b878e688df2f54f3899560ba72d2992042af.

+++ drivers/staging/comedi/drivers/das800.c
@@ -427,7 +427,7 @@ static irqreturn_t das800_interrupt(int irq, void *d)
 	struct comedi_cmd *cmd;
 	unsigned long irq_flags;
 	unsigned int status;
-	unsigned short val;
+	unsigned int val;
 	bool fifo_empty;
 	bool fifo_overflow;
 	int i;
